Reorganise plotting
