+++ gwemopt/io/__init__.py
@@ -1,0 +1,5 @@
+from gwemopt.io.export_tiles import (
+    export_tiles_coverage_hist,
+    export_tiles_coverage_int,
+)
+from gwemopt.io.schedule import read_schedule
