+++ gwemopt/io/schedule.py
