+++ gwemopt/run.py
@@ -13,6 +13,13 @@ from gwemopt.args import parse_args
 from gwemopt.gracedb import get_event
 from gwemopt.params import params_struct
 from gwemopt.paths import DEFAULT_BASE_OUTPUT_DIR
+from gwemopt.plotting import (
+    make_coverage_plots,
+    make_efficiency_plots,
+    make_tile_plots,
+    plot_observability,
+    plot_skymap,
+)
 from gwemopt.utils import read_skymap
 
 
@@ -60,7 +67,7 @@ def run(args):
 
     if args.doPlots:
         print("Plotting skymap...")
-        gwemopt.plotting.skymap(params, map_struct)
+        plot_skymap(params, map_struct)
 
     if args.doObservability:
         print("Generating observability")
@@ -68,7 +75,7 @@ def run(args):
         map_struct["observability"] = observability_struct
         if args.doPlots:
             print("Plotting observability...")
-            gwemopt.plotting.observability(params, map_struct)
+            plot_observability(params, map_struct)
         if args.doObservabilityExit:
             for telescope in params["telescopes"]:
                 if (
@@ -128,7 +135,7 @@ def run(args):
 
         if args.doPlots:
             print("Plotting tiles struct...")
-            gwemopt.plotting.tiles(params, map_struct, tile_structs)
+            make_tile_plots(params, map_struct, tile_structs)
 
     if args.doSchedule:
         if args.doTiles:
@@ -157,11 +164,11 @@ def run(args):
         if args.doPlots:
             print("Plotting coverage...")
             if args.doCatalog:
-                gwemopt.plotting.coverage(
+                make_coverage_plots(
                     params, map_struct, coverage_struct, catalog_struct=catalog_struct
                 )
             else:
-                gwemopt.plotting.coverage(params, map_struct, coverage_struct)
+                make_coverage_plots(params, map_struct, coverage_struct)
 
     if args.doEfficiency:
         if args.doSchedule or args.doCoverage:
@@ -195,7 +202,7 @@ def run(args):
 
             if args.doPlots:
                 print("Plotting efficiency...")
-                gwemopt.plotting.efficiency(params, map_struct, efficiency_structs)
+                make_efficiency_plots(params, map_struct, efficiency_structs)
         else:
             print("Need to enable --doSchedule or --doCoverage for --doEfficiency")
             exit(0)
